+++ src/client.ts
@@ -45,6 +45,11 @@ export class Client {
     return result as { path: string }
   }
 
+  async fileCreateFolder(props: { name: string; folder?: string }) {
+    const result = await this.request('/file/create-folder', props)
+    return result as { path: string }
+  }
+
   async fileDelete(props: { path: string }) {
     const result = await this.request('/file/delete', props)
     return result as { path: string }
@@ -63,28 +68,6 @@ export class Client {
   async fileRead(props: { path: string }) {
     const result = await this.request('/file/read', props)
     return result as { bytes: ArrayBuffer }
-  }
-
-  // Folder
-
-  async folderCopy(props: { path: string; folder?: string }) {
-    const result = await this.request('/folder/copy', props)
-    return result as { path: string }
-  }
-
-  async folderCreate(props: { name: string; folder?: string }) {
-    const result = await this.request('/folder/create', props)
-    return result as { path: string }
-  }
-
-  async folderDelete(props: { path: string }) {
-    const result = await this.request('/folder/delete', props)
-    return result as { path: string }
-  }
-
-  async folderMove(props: { path: string; folder: string }) {
-    const result = await this.request('/folder/move', props)
-    return result as { path: string }
   }
 
   // Package
