--- src/components/Controllers/Files/Dialogs/TargetDialog.tsx
@@ -1,90 +1,0 @@
-import * as React from 'react'
-import capitalize from 'lodash/capitalize'
-import Dialog from '@mui/material/Dialog'
-import DialogTitle from '@mui/material/DialogTitle'
-import IconButton from '@mui/material/IconButton'
-import Button from '@mui/material/Button'
-import CloseIcon from '@mui/icons-material/Close'
-import Box from '@mui/material/Box'
-import Columns from '../../../Parts/Columns'
-import { DialogContent } from '@mui/material'
-import FileTree from '../../../Parts/Trees/FileTree'
-import { useStore, selectors } from '../store'
-
-export default function TargetDialog() {
-  const [target, setTarget] = React.useState('')
-  const targetTree = useStore(selectors.targetTree)
-  const dialog = useStore((state) => state.dialog)
-  const setDialog = useStore((state) => state.setDialog)
-  const copyFile = useStore((state) => state.copyFile)
-  const moveFile = useStore((state) => state.moveFile)
-  const copyFolder = useStore((state) => state.copyFolder)
-  const moveFolder = useStore((state) => state.moveFolder)
-  const isFolder = useStore(selectors.isFolder)
-  const handleClose = () => setDialog(undefined)
-  const handleSelect = () => {
-    const copyAction = isFolder ? copyFolder : copyFile
-    const moveAction = isFolder ? moveFolder : moveFile
-    const action = dialog === 'copy' ? copyAction : moveAction
-    action(target)
-    setDialog(undefined)
-  }
-  return (
-    <Dialog
-      fullWidth
-      maxWidth="sm"
-      onClose={handleClose}
-      open={dialog === 'copy' || dialog === 'move'}
-    >
-      <DialogTitle>
-        {capitalize(dialog)} {isFolder ? 'Folder' : 'File'}
-        <IconButton
-          aria-label="close"
-          onClick={handleClose}
-          sx={{ position: 'absolute', right: 8, top: 8, color: 'grey' }}
-        >
-          <CloseIcon />
-        </IconButton>
-      </DialogTitle>
-      <DialogContent
-        sx={{
-          borderStyle: 'dotted',
-          borderWidth: 1,
-          borderRadius: 1,
-          marginLeft: 3,
-          marginRight: 3,
-        }}
-      >
-        <FileTree
-          tree={targetTree}
-          expanded={[targetTree[0].path]}
-          onPathChange={setTarget}
-        />
-      </DialogContent>
-      <Box sx={{ paddingX: 3, paddingY: 1 }}>
-        <Columns spacing={2}>
-          <Button
-            fullWidth
-            sx={{ my: 0.5 }}
-            variant="contained"
-            size="small"
-            onClick={handleClose}
-            color="warning"
-          >
-            Cancel
-          </Button>
-          <Button
-            fullWidth
-            sx={{ my: 0.5 }}
-            variant="contained"
-            size="small"
-            onClick={handleSelect}
-            color="secondary"
-          >
-            {dialog === 'copy' ? 'Copy' : 'Move'}
-          </Button>
-        </Columns>
-      </Box>
-    </Dialog>
-  )
-}
